requirements.txt
PyYAML==6.0
requests==2.27.1
python-dateutil==2.8.2
